some issues fixed
