+++ asm/asm_dasha.h
@@ -39,5 +39,13 @@ _Bool				token_sequence(t_token *token);
 u_int32_t			calc_command_weight(t_token *token);
 
 void				print_header(int fd, header_t *header);
+void				if_label(t_token *token, t_token_sec *check_list);
+void				if_arg(t_token *token, t_token_sec *check_list);
+void				if_name(t_token *token, t_token_sec *check_list);
+void				if_nl(t_token *token, t_token_sec *check_list);
+void				if_comment_prog(t_token *token, t_token_sec *check_list);
+void				command_check(t_token *token);
+void				init_headers(header_t *header, t_token *token);
+
 
 #endif
